WCM-112 Fix plugin names
