+++ rule-device-test/build.xml
@@ -3,7 +3,7 @@
 <project name="rule-device-test" basedir="." default="compile">
 	<property name="plugin.version" value="1" />
 
-	<property name="plugin.name" value="com.liferay.portal.devicerule.test" />
+	<property name="plugin.name" value="com.liferay.content.targeting.rule.device.test" />
 
 	<property name="import.shared" value="rule-device,service-test-util" />
