+++ rule-organization-member-test/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Organization Member Test
+Bundle-Name: Organization Member Rule Test
 Bundle-SymbolicName: com.liferay.content.targeting.rule.organization.member.test
 Bundle-Version: 1.0.0
 Include-Resource: classes
