+++ rule-organization-member-test/build.xml
@@ -3,7 +3,7 @@
 <project name="rule-organization-member-test" basedir="." default="compile">
 	<property name="plugin.version" value="1" />
 
-	<property name="plugin.name" value="com.liferay.portal.organization.member.test" />
+	<property name="plugin.name" value="com.liferay.content.targeting.rule.organization.member.test" />
 
 	<property name="import.shared" value="rule-organization-member,service-test-util" />
