+++ rule-organization-member-test/test/integration/com/liferay/content/targeting/rule/organization/member/OrganizationMemberRuleTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.content.targeting.rule.role;
+package com.liferay.content.targeting.rule.organization.member;
 
 import com.liferay.content.targeting.anonymous.users.model.AnonymousUser;
 import com.liferay.content.targeting.anonymous.users.service.AnonymousUserLocalService;
@@ -23,22 +23,12 @@ import com.liferay.content.targeting.service.RuleInstanceLocalService;
 import com.liferay.content.targeting.service.test.service.ServiceTestUtil;
 import com.liferay.content.targeting.service.test.util.TestPropsValues;
 import com.liferay.osgi.util.service.ServiceTrackerUtil;
-import com.liferay.portal.kernel.exception.PortalException;
-import com.liferay.portal.kernel.exception.SystemException;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
-import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.model.Organization;
 import com.liferay.portal.model.OrganizationConstants;
-import com.liferay.portal.model.Role;
-import com.liferay.portal.model.RoleConstants;
 import com.liferay.portal.service.OrganizationLocalServiceUtil;
-import com.liferay.portal.service.RoleLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.UserGroupRoleLocalServiceUtil;
-
-import java.util.List;
 
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
@@ -55,7 +45,7 @@ import org.osgi.framework.BundleException;
  * @author Eudaldo Alonso
  */
 @RunWith(Arquillian.class)
-public class OrganizationRoleTest {
+public class OrganizationMemberRuleTest {
 
 	@Before
 	public void setUp() {
@@ -75,7 +65,7 @@ public class OrganizationRoleTest {
 	}
 
 	@Test
-	public void testOrganizationRoleRule() throws Exception {
+	public void testOrganizationMemberRule() throws Exception {
 		ServiceContext serviceContext = ServiceTestUtil.getServiceContext();
 
 		AnonymousUser anonymousUser =
@@ -83,42 +73,22 @@ public class OrganizationRoleTest {
 				TestPropsValues.getUserId(), "127.0.0.1", StringPool.BLANK,
 				serviceContext);
 
-		List<Role> roles = RoleLocalServiceUtil.getRoles(
-			TestPropsValues.getCompanyId(),
-			new int[]{RoleConstants.TYPE_ORGANIZATION});
+		Rule rule = _rulesRegistry.getRule("OrganizationMemberRule");
 
-		Role role = roles.get(0);
+		Organization organization =
+			OrganizationLocalServiceUtil.addOrganization(
+				TestPropsValues.getUserId(),
+				OrganizationConstants.DEFAULT_PARENT_ORGANIZATION_ID,
+				StringUtil.randomString(), false);
 
-		Rule rule = _rulesRegistry.getRule("OrganizationRoleRule");
+		OrganizationLocalServiceUtil.addUserOrganization(
+			TestPropsValues.getUserId(), organization.getOrganizationId());
 
 		RuleInstance ruleInstance = _ruleInstanceLocalService.addRuleInstance(
 			TestPropsValues.getUserId(), rule.getRuleKey(), 0,
-			getTypeSettings(role.getRoleId()), serviceContext);
-
-		UserGroupRoleLocalServiceUtil.addUserGroupRoles(
-			new long[]{TestPropsValues.getUserId()}, _organization.getGroupId(),
-			role.getRoleId());
+			String.valueOf(organization.getOrganizationId()), serviceContext);
 
 		Assert.assertTrue(rule.evaluate(null, ruleInstance, anonymousUser));
-	}
-
-	protected String getTypeSettings(long roleId)
-		throws PortalException, SystemException {
-
-		_organization = OrganizationLocalServiceUtil.addOrganization(
-			TestPropsValues.getUserId(),
-			OrganizationConstants.DEFAULT_PARENT_ORGANIZATION_ID,
-			StringUtil.randomString(), false);
-
-		OrganizationLocalServiceUtil.addUserOrganization(
-			TestPropsValues.getUserId(), _organization.getOrganizationId());
-
-		JSONObject jsonObj = JSONFactoryUtil.createJSONObject();
-
-		jsonObj.put("organizationId", _organization.getOrganizationId());
-		jsonObj.put("roleId", roleId);
-
-		return jsonObj.toString();
 	}
 
 	private AnonymousUserLocalService _anonymousUserLocalService;
@@ -126,7 +96,6 @@ public class OrganizationRoleTest {
 	@ArquillianResource
 	private Bundle _bundle;
 
-	private Organization _organization;
 	private RuleInstanceLocalService _ruleInstanceLocalService;
 	private RulesRegistry _rulesRegistry;
