+++ rule-role-test/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Rule Role Test
+Bundle-Name: Role Rule Test
 Bundle-SymbolicName: com.liferay.content.targeting.rule.role.test
 Bundle-Version: 1.0.0
 Include-Resource: classes
