+++ rule-role-test/test/integration/com/liferay/content/targeting/rule/role/RegularRoleRuleTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.content.targeting.rule.user.group.member;
+package com.liferay.content.targeting.rule.role;
 
 import com.liferay.content.targeting.anonymous.users.model.AnonymousUser;
 import com.liferay.content.targeting.anonymous.users.service.AnonymousUserLocalService;
@@ -24,10 +24,12 @@ import com.liferay.content.targeting.service.test.service.ServiceTestUtil;
 import com.liferay.content.targeting.service.test.util.TestPropsValues;
 import com.liferay.osgi.util.service.ServiceTrackerUtil;
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.model.UserGroup;
+import com.liferay.portal.model.Role;
+import com.liferay.portal.model.RoleConstants;
+import com.liferay.portal.service.RoleLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.UserGroupLocalServiceUtil;
+
+import java.util.List;
 
 import org.jboss.arquillian.junit.Arquillian;
 import org.jboss.arquillian.test.api.ArquillianResource;
@@ -44,7 +46,7 @@ import org.osgi.framework.BundleException;
  * @author Eudaldo Alonso
  */
 @RunWith(Arquillian.class)
-public class UserGroupMemberTest {
+public class RegularRoleRuleTest {
 
 	@Before
 	public void setUp() {
@@ -64,7 +66,7 @@ public class UserGroupMemberTest {
 	}
 
 	@Test
-	public void testUserGroupMemberRule() throws Exception {
+	public void testRegularRoleRule() throws Exception {
 		ServiceContext serviceContext = ServiceTestUtil.getServiceContext();
 
 		AnonymousUser anonymousUser =
@@ -72,20 +74,20 @@ public class UserGroupMemberTest {
 				TestPropsValues.getUserId(), "127.0.0.1", StringPool.BLANK,
 				serviceContext);
 
-		Rule rule = _rulesRegistry.getRule("UserGroupMemberRule");
+		List<Role> roles = RoleLocalServiceUtil.getRoles(
+			TestPropsValues.getCompanyId(),
+			new int[] {RoleConstants.TYPE_REGULAR});
 
-		UserGroup userGroup =
-			UserGroupLocalServiceUtil.addUserGroup(
-				TestPropsValues.getUserId(), TestPropsValues.getCompanyId(),
-				StringUtil.randomString(), StringUtil.randomString(),
-				new ServiceContext());
+		Role role = roles.get(0);
 
-		UserGroupLocalServiceUtil.addUserUserGroup(
-			TestPropsValues.getUserId(), userGroup.getUserGroupId());
+		Rule rule = _rulesRegistry.getRule("RegularRoleRule");
 
 		RuleInstance ruleInstance = _ruleInstanceLocalService.addRuleInstance(
 			TestPropsValues.getUserId(), rule.getRuleKey(), 0,
-			String.valueOf(userGroup.getUserGroupId()), serviceContext);
+			String.valueOf(role.getRoleId()), serviceContext);
+
+		RoleLocalServiceUtil.addUserRole(
+			TestPropsValues.getUserId(), role.getRoleId());
 
 		Assert.assertTrue(rule.evaluate(null, ruleInstance, anonymousUser));
 	}
