+++ rule-site-member-test/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Site Member Test
+Bundle-Name: Site Member Rule Test
 Bundle-SymbolicName: com.liferay.content.targeting.rule.site.member.test
 Bundle-Version: 1.0.0
 Include-Resource: classes
