+++ rule-user-group-member-test/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: UserGroup Member Test
+Bundle-Name: User Group Member Rule Test
 Bundle-SymbolicName: com.liferay.content.targeting.rule.user.group.member.test
 Bundle-Version: 1.0.0
 Include-Resource: classes
