+++ rule-user-group-member/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: UserGroup Member Rule
+Bundle-Name: User Group Member Rule
 Bundle-SymbolicName: com.liferay.content.targeting.rule.user.group.member
 Bundle-Version: 1.0.0
 Import-Package:\
