+++ rule-user-logged-test/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: User Logged Test
+Bundle-Name: User Logged Rule Test
 Bundle-SymbolicName: com.liferay.content.targeting.rule.user.logged.test
 Bundle-Version: 1.0.0
 Include-Resource: classes
