+++ rule-visited-test/bnd.bnd
@@ -1,4 +1,4 @@
-Bundle-Name: Visited Test
+Bundle-Name: Visited Rule Test
 Bundle-SymbolicName: com.liferay.content.targeting.rule.visited.test
 Bundle-Version: 1.0.0
 Include-Resource: classes
