+++ rule-visited-test/test/integration/com/liferay/content/targeting/rule/visited/PageVisitedRuleTest.java
@@ -12,8 +12,9 @@
  * details.
  */
 
-package com.liferay.content.targeting.rule.organization.member;
+package com.liferay.content.targeting.rule.visited;
 
+import com.liferay.content.targeting.analytics.service.AnalyticsEventLocalService;
 import com.liferay.content.targeting.anonymous.users.model.AnonymousUser;
 import com.liferay.content.targeting.anonymous.users.service.AnonymousUserLocalService;
 import com.liferay.content.targeting.api.model.Rule;
@@ -21,13 +22,14 @@ import com.liferay.content.targeting.api.model.RulesRegistry;
 import com.liferay.content.targeting.model.RuleInstance;
 import com.liferay.content.targeting.service.RuleInstanceLocalService;
 import com.liferay.content.targeting.service.test.service.ServiceTestUtil;
+import com.liferay.content.targeting.service.test.util.GroupTestUtil;
 import com.liferay.content.targeting.service.test.util.TestPropsValues;
 import com.liferay.osgi.util.service.ServiceTrackerUtil;
 import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.model.Organization;
-import com.liferay.portal.model.OrganizationConstants;
-import com.liferay.portal.service.OrganizationLocalServiceUtil;
+import com.liferay.portal.model.Group;
+import com.liferay.portal.model.Layout;
+import com.liferay.portal.model.LayoutConstants;
+import com.liferay.portal.service.LayoutLocalServiceUtil;
 import com.liferay.portal.service.ServiceContext;
 
 import org.jboss.arquillian.junit.Arquillian;
@@ -45,7 +47,7 @@ import org.osgi.framework.BundleException;
  * @author Eudaldo Alonso
  */
 @RunWith(Arquillian.class)
-public class OrganizationMemberTest {
+public class PageVisitedRuleTest {
 
 	@Before
 	public void setUp() {
@@ -56,6 +58,8 @@ public class OrganizationMemberTest {
 			e.printStackTrace();
 		}
 
+		_analyticsEventLocalService = ServiceTrackerUtil.getService(
+			AnalyticsEventLocalService.class, _bundle.getBundleContext());
 		_anonymousUserLocalService = ServiceTrackerUtil.getService(
 			AnonymousUserLocalService.class, _bundle.getBundleContext());
 		_ruleInstanceLocalService = ServiceTrackerUtil.getService(
@@ -65,32 +69,66 @@ public class OrganizationMemberTest {
 	}
 
 	@Test
-	public void testOrganizationMemberRule() throws Exception {
-		ServiceContext serviceContext = ServiceTestUtil.getServiceContext();
+	public void testNotVisitedPageRule() throws Exception {
+		Group group = GroupTestUtil.addGroup();
+
+		ServiceContext serviceContext = ServiceTestUtil.getServiceContext(
+			group.getGroupId(), TestPropsValues.getUserId());
 
 		AnonymousUser anonymousUser =
 			_anonymousUserLocalService.addAnonymousUser(
 				TestPropsValues.getUserId(), "127.0.0.1", StringPool.BLANK,
 				serviceContext);
 
-		Rule rule = _rulesRegistry.getRule("OrganizationMemberRule");
+		Rule rule = _rulesRegistry.getRule("PageVisitedRule");
 
-		Organization organization =
-			OrganizationLocalServiceUtil.addOrganization(
-				TestPropsValues.getUserId(),
-				OrganizationConstants.DEFAULT_PARENT_ORGANIZATION_ID,
-				StringUtil.randomString(), false);
-
-		OrganizationLocalServiceUtil.addUserOrganization(
-			TestPropsValues.getUserId(), organization.getOrganizationId());
+		Layout layout = LayoutLocalServiceUtil.addLayout(
+			TestPropsValues.getUserId(), group.getGroupId(), false,
+			LayoutConstants.DEFAULT_PARENT_LAYOUT_ID, "Layout", "Layout",
+			StringPool.BLANK, LayoutConstants.TYPE_PORTLET, false,
+			StringPool.BLANK, serviceContext);
 
 		RuleInstance ruleInstance = _ruleInstanceLocalService.addRuleInstance(
 			TestPropsValues.getUserId(), rule.getRuleKey(), 0,
-			String.valueOf(organization.getOrganizationId()), serviceContext);
+			String.valueOf(layout.getPlid()), serviceContext);
+
+		Assert.assertFalse(rule.evaluate(null, ruleInstance, anonymousUser));
+	}
+
+	@Test
+	public void testVisitedPageRule() throws Exception {
+		Group group = GroupTestUtil.addGroup();
+
+		ServiceContext serviceContext = ServiceTestUtil.getServiceContext(
+			group.getGroupId(), TestPropsValues.getUserId());
+
+		AnonymousUser anonymousUser =
+			_anonymousUserLocalService.addAnonymousUser(
+				TestPropsValues.getUserId(), "127.0.0.1", StringPool.BLANK,
+				serviceContext);
+
+		Rule rule = _rulesRegistry.getRule("PageVisitedRule");
+
+		Layout layout = LayoutLocalServiceUtil.addLayout(
+			TestPropsValues.getUserId(), group.getGroupId(), false,
+			LayoutConstants.DEFAULT_PARENT_LAYOUT_ID, "Layout", "Layout",
+			StringPool.BLANK, LayoutConstants.TYPE_PORTLET, false,
+			StringPool.BLANK, serviceContext);
+
+		_analyticsEventLocalService.addAnalyticsEvent(
+			1, anonymousUser.getAnonymousUserId(), Layout.class.getName(),
+			layout.getPlid(), Layout.class.getName(), new long[]{1}, null,
+			"view", "127.0.0.1", "User Agent", "ES", "http://localhost", null,
+			serviceContext);
+
+		RuleInstance ruleInstance = _ruleInstanceLocalService.addRuleInstance(
+			TestPropsValues.getUserId(), rule.getRuleKey(), 0,
+			String.valueOf(layout.getPlid()), serviceContext);
 
 		Assert.assertTrue(rule.evaluate(null, ruleInstance, anonymousUser));
 	}
 
+	private AnalyticsEventLocalService _analyticsEventLocalService;
 	private AnonymousUserLocalService _anonymousUserLocalService;
 
 	@ArquillianResource
